add timings
