+++ posthog/api/element.py
@@ -12,6 +12,7 @@ from posthog.models.element.element import chain_to_elements
 from posthog.models.element.sql import GET_ELEMENTS, GET_VALUES
 from posthog.models.instance_setting import get_instance_setting
 from posthog.models.property.util import parse_prop_grouped_clauses
+from posthog.models.utils import ServerTimingsGathered
 from posthog.queries.query_date_range import QueryDateRange
 from posthog.utils import format_query_params_absolute_url
 
@@ -49,72 +50,84 @@ class ElementViewSet(TeamAndOrgViewSetMixin, viewsets.ModelViewSet):
         Currently only $autocapture and $rageclick are supported
         """
 
-        sample_rows_count = get_instance_setting("HEATMAP_SAMPLE_N") or 2_000_000
+        timer = ServerTimingsGathered()
 
-        filter = Filter(request=request, team=self.team)
+        with timer("prepare_for_query"):
+            sample_rows_count = get_instance_setting("HEATMAP_SAMPLE_N") or 2_000_000
 
-        date_params = {}
-        query_date_range = QueryDateRange(filter=filter, team=self.team, should_round=True)
-        date_from, date_from_params = query_date_range.date_from
-        date_to, date_to_params = query_date_range.date_to
-        date_params.update(date_from_params)
-        date_params.update(date_to_params)
+            filter = Filter(request=request, team=self.team)
 
-        try:
-            limit = int(request.query_params.get("limit", 250))
-        except ValueError:
-            raise ValidationError("Limit must be an integer")
+            date_params = {}
+            query_date_range = QueryDateRange(filter=filter, team=self.team, should_round=True)
+            date_from, date_from_params = query_date_range.date_from
+            date_to, date_to_params = query_date_range.date_to
+            date_params.update(date_from_params)
+            date_params.update(date_to_params)
 
-        try:
-            offset = int(request.query_params.get("offset", 0))
-        except ValueError:
-            raise ValidationError("offset must be an integer")
+            try:
+                limit = int(request.query_params.get("limit", 250))
+            except ValueError:
+                raise ValidationError("Limit must be an integer")
 
-        events_filter = self._events_filter(request)
+            try:
+                offset = int(request.query_params.get("offset", 0))
+            except ValueError:
+                raise ValidationError("offset must be an integer")
 
-        prop_filters, prop_filter_params = parse_prop_grouped_clauses(
-            team_id=self.team.pk,
-            property_group=filter.property_groups,
-            hogql_context=filter.hogql_context,
-        )
-        result = sync_execute(
-            GET_ELEMENTS.format(
-                date_from=date_from,
-                date_to=date_to,
-                query=prop_filters,
-                limit=limit + 1,
-                offset=offset,
-            ),
-            {
-                "team_id": self.team.pk,
-                "timezone": self.team.timezone,
-                "sample_rows_count": sample_rows_count,
-                **prop_filter_params,
-                **date_params,
-                "filter_event_types": events_filter,
-                **filter.hogql_context.values,
-            },
-        )
-        serialized_elements = [
-            {
-                "count": elements[1],
-                "hash": None,
-                "type": elements[2],
-                "elements": [ElementSerializer(element).data for element in chain_to_elements(elements[0])],
-            }
-            for elements in result[:limit]
-        ]
+            events_filter = self._events_filter(request)
+
+            # unless someone is using this as an API client, this is only for the toolbar,
+            # which only ever queries date range, event type, and URL
+            prop_filters, prop_filter_params = parse_prop_grouped_clauses(
+                team_id=self.team.pk,
+                property_group=filter.property_groups,
+                hogql_context=filter.hogql_context,
+            )
+
+        with timer("execute_query"):
+            result = sync_execute(
+                GET_ELEMENTS.format(
+                    date_from=date_from,
+                    date_to=date_to,
+                    query=prop_filters,
+                    limit=limit + 1,
+                    offset=offset,
+                ),
+                {
+                    "team_id": self.team.pk,
+                    "timezone": self.team.timezone,
+                    "sample_rows_count": sample_rows_count,
+                    **prop_filter_params,
+                    **date_params,
+                    "filter_event_types": events_filter,
+                    **filter.hogql_context.values,
+                },
+            )
+
+        with timer("serialize_elements"):
+            serialized_elements = [
+                {
+                    "count": elements[1],
+                    "hash": None,
+                    "type": elements[2],
+                    "elements": [ElementSerializer(element).data for element in chain_to_elements(elements[0])],
+                }
+                for elements in result[:limit]
+            ]
 
         has_next = len(result) == limit + 1
         next_url = format_query_params_absolute_url(request, offset + limit) if has_next else None
         previous_url = format_query_params_absolute_url(request, offset - limit) if offset - limit >= 0 else None
-        return response.Response(
+        elements_response = response.Response(
             {
                 "results": serialized_elements,
                 "next": next_url,
                 "previous": previous_url,
             }
         )
+
+        elements_response.headers["Server-Timing"] = timer.to_header_string()
+        return elements_response
 
     def _events_filter(self, request) -> tuple[Literal["$autocapture", "$rageclick", "$dead_click"], ...]:
         SUPPORTED_EVENTS = {"$autocapture", "$rageclick", "$dead_click"}
