+++ posthog/models/utils.py
@@ -17,6 +17,7 @@ from django.db.models.constraints import BaseConstraint
 from django.utils.text import slugify
 
 from posthog.constants import MAX_SLUG_LENGTH
+from posthog.schema import QueryTiming
 
 if TYPE_CHECKING:
     from random import Random
@@ -401,3 +402,45 @@ class RootTeamMixin(models.Model):
         if hasattr(self, "team") and self.team and hasattr(self.team, "parent_team") and self.team.parent_team:  # type: ignore
             self.team = self.team.parent_team  # type: ignore
         super().save(*args, **kwargs)
+
+
+# context manager for gathering a sequence of server timings
+# can be used to then return the timings in the HTTP response headers
+# so that browsers and other tools can show them
+class ServerTimingsGathered:
+    def __init__(self):
+        # Instance level dictionary to store timings
+        self.timings_dict = {}
+
+    def __call__(self, name):
+        self.name = name
+        return self
+
+    def __enter__(self):
+        # timings are assumed to be in milliseconds when reported
+        # but are gathered by time.perf_counter which is fractional seconds 🫠
+        # so each value is multiplied by 1000 at collection
+        self.start_time = time.perf_counter() * 1000
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        end_time = time.perf_counter() * 1000
+        elapsed_time = end_time - self.start_time
+        self.timings_dict[self.name] = elapsed_time
+
+    def get_all_timings(self):
+        return self.timings_dict
+
+    def generate_timings(self, hogql_timings: list[QueryTiming] | None) -> dict[str, float]:
+        timings_dict = self.get_all_timings()
+        hogql_timings_dict = {}
+        for key, value in hogql_timings or {}:
+            new_key = f"hogql_{key[1].lstrip('./').replace('/', '_')}"
+            # HogQL query timings are in seconds, convert to milliseconds
+            hogql_timings_dict[new_key] = value[1] * 1000
+        all_timings = {**timings_dict, **hogql_timings_dict}
+        return all_timings
+
+    def to_header_string(self, hogql_timings: list[QueryTiming] | None):
+        return ", ".join(
+            f"{key};dur={round(duration, ndigits=2)}" for key, duration in self.generate_timings(hogql_timings).items()
+        )
